ref:スクリプトのMVPを一部編集
